feat/#104: 동네 모달 분리 및 동네 선택 기능 추가
